@@ -1,10 +1,10 @@
 name: giphy_picker
 description: A plugin that allows you to pick animated GIF images from Giphy.
-version: 1.0.8
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/firstfloorsoftware/giphy_picker
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
